Initial commit.
So far:
The app consists of a single page containing a map of all of the BART
stations. In this commit, there are boxes around all of the stations on
the Richmond (red) line. Each box is tied to a click handler. Clicking
on a station's box queries the BART API for that station's current
estimated departure times. The name of the station is read from the
query result and displayed in the console.